pyproject.toml
[project]
name = "photo-scanner"
version = "0.1.0"
description = ""
authors = [
    {name = "Arjav Jain",email = "jarjav2005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "tensorflow (>=2.20.0,<3.0.0)",
    "numpy (>=2.4.1,<3.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "ftfy (>=6.3.1,<7.0.0)",
    "regex (>=2026.1.15,<2027.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sentence-transformers (>=5.2.2,<6.0.0)",
    "easyocr (>=1.7.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "optimum[onnxruntime] (>=2.1.0,<3.0.0)",
    "onnxruntime (>=1.23.2,<2.0.0)",
    "onnx (>=1.20.1,<2.0.0)",
    "tf-keras (>=2.20.1,<3.0.0)",
    "psutil (>=7.2.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
